pyproject.toml
[project]
classifiers = [
    "License :: OSI Approved :: All rights reserved (c) Israel Llorens",
]
name = "Framework"
version = "0.1.0"
description = "Librería común para desarrollo Spark y soporte al mismo para Ingeniería de Datos y Ciencia"
authors = [{"name" = "Israel Llorens", "email" = "sanchezis@hotmail.com"}]
license = "All rights reserved (c) Israel Llorens"  # SPDX expression
readme = "README.md"
requires-python = ">=3.11,<3.14"


dependencies = [
    "pyspark==3.4.4",
    "networkx>=3.4.2",
    "seaborn>=0.13.2",
    "plotly>=6.0.1",
    "statsmodels>=0.14.4",
    "scipy>=1.15.2",
    "ipykernel>=6.29.5",
    "scikit-learn>=1.6.1",
    "certifi>=2025.1.31",
    "numpy (==1.26.4)",
    "requests (>=2.32.3,<3.0.0)",
    "beautifulsoup4 (>=4.13.3,<5.0.0)",
    "bs4 (>=0.0.2,<0.0.3)",
    "graphframes (>=0.6,<0.7)",
    "pm4py (>=2.7.14.4,<3.0.0.0)",
    "pytest (>=8.3.5,<9.0.0)",
    "nbformat (>=4.2.0)",
    "pgmpy (>=1.0.0,<2.0.0)",
    "llama-stack (>=0.2.2,<0.3.0)",

    "protobuf (>=6.30.2,<7.0.0)",
    "pyarrow (==12.0.1)",
    "py4j (==0.10.9.7)",

    "openllm (>=0.6.30,<0.7.0)",
    "spark-nlp (==5.2.0)",
    "spacy (==3.7.0)",
    "en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.0/en_core_web_sm-3.7.0-py3-none-any.whl",

    "faiss-cpu==1.7.4",
    "torch==2.0.1",
    "torchaudio==2.0.2",
    "torchvision==0.15.2",
    "accelerate==0.20.3",
    "transformers==4.40.0",
    "sentencepiece==0.2.0",
    "sentence-transformers==2.7.0",

    "onnxruntime-silicon>=1.16.3,<2.0.0",

    "openai==1.73.0",

    "flask>=3.1.1,<4.0.0",
    "flask-cors>=6.0.0,<7.0.0",
    "streamlit>=1.45.1,<2.0.0",
    "tornado>=6.5.1,<7.0.0",
    "python-dotenv>=1.1.0,<2.0.0",

    # 🧠 Redes y grafos
    "gremlinpython>=3.7.3,<4.0.0",

    "ipywidgets>=8.1.7,<9.0.0",
    "chardet>=5.2.0,<6.0.0",

    "kafka-python>=2.2.10,<3.0.0",

    "aiohttp (>=3.12.6,<4.0.0)",

    "pinecone (>=7.0.2,<8.0.0)",

]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
transformers = "4.40.0"
huggingface-hub = "0.14.1"